Refactor of fio.py into fio and struct submodules
