+++ src/fiat/models/geom.py
@@ -132,7 +132,7 @@ class GeomModel(BaseModel):
         """Set the chunking size."""
         # Determine maximum geometry dataset size
         max_geom_size = max(
-            [item.size for item in self.exposure_geoms.values()],
+            [item.layer.size for item in self.exposure_geoms.values()],
         )
         # Set the 1D chunks
         self.chunks = create_1d_chunk(
@@ -162,9 +162,11 @@ class GeomModel(BaseModel):
             with open_geom(
                 Path(self.cfg.get("output.path"), out_geom), mode="w", overwrite=True
             ) as _w:
-                _w.create_layer(self.srs, gm.geom_type)
-                _w.create_fields(dict(zip(gm.fields, gm.dtypes)))
-                _w.create_fields(dict(zip(new_fields, [ogr.OFTReal] * len(new_fields))))
+                _w.create_layer(self.srs, gm.layer.geom_type)
+                _w.layer.create_fields(dict(zip(gm.layer.fields, gm.layer.dtypes)))
+                _w.layer.create_fields(
+                    dict(zip(new_fields, [ogr.OFTReal] * len(new_fields)))
+                )
             _w = None
 
             # Check whether to do the same for the csv
@@ -194,7 +196,7 @@ class GeomModel(BaseModel):
                 self.cfg.get("exposure.csv.settings.index"),
             )
         for key, gm in self.exposure_geoms.items():
-            columns = gm._columns
+            columns = gm.layer._columns
             self._discover_exposure_meta(
                 columns,
                 meta,
@@ -306,15 +308,15 @@ class GeomModel(BaseModel):
 
             # check the internal srs of the file
             check_internal_srs(
-                data.srs,
+                data.layer.srs,
                 path.name,
             )
 
             # check if file srs is the same as the model srs
-            if not check_vs_srs(self.srs, data.srs):
+            if not check_vs_srs(self.srs, data.layer.srs):
                 logger.warning(
                     f"Spatial reference of '{path.name}' \
-('{get_srs_repr(data.srs)}') does not match \
+('{get_srs_repr(data.layer.srs)}') does not match \
 the model spatial reference ('{get_srs_repr(self.srs)}')"
                 )
                 logger.info(f"Reprojecting '{path.name}' to '{get_srs_repr(self.srs)}'")
@@ -322,7 +324,7 @@ the model spatial reference ('{get_srs_repr(self.srs)}')"
 
             # check if it falls within the extent of the hazard map
             check_geom_extent(
-                data.bounds,
+                data.layer.bounds,
                 self.hazard_grid.bounds,
             )
