+++ src/fiat/struct/__init__.py
@@ -1,0 +1,4 @@
+"""Data structures (glorified classes)."""
+
+from .geom import *
+from .table import *
