+++ src/fiat/struct/grid.py
@@ -1,0 +1,1 @@
+"""A python wrapper structure for gdal Band."""
